corrected test_utils
