+++ tests/test_libcli.py
@@ -57,7 +57,14 @@ def test_KwargsDict_from_file(cmd, expected):
     assert args.d == expected
 
 
-def test_KwargsDict_error():
+@pytest.mark.parametrize(
+    's,error',
+    [
+        ("-d {fail}", JSONDecodeError),
+        ("-d fail", FileNotFoundError),
+        ]
+    )
+def test_KwargsDict_error(s, error):
     """Test error."""
     ap = argparse.ArgumentParser()
     ap.add_argument(
@@ -65,8 +72,8 @@ def test_KwargsDict_error():
         action=libcli.KwargsDict,
         default=None,
         )
-    with pytest.raises(JSONDecodeError):
-        ap.parse_args("-d fail".split())
+    with pytest.raises(error):
+        ap.parse_args(s.split())
 
 
 def test_find_AnalysisBase_members():
